addresses feedback
