+++ elfpy/eth/__init__.py
@@ -3,6 +3,6 @@
 from . import abi, accounts
 from .numeric_utils import convert_scaled_value
 from .receipts import get_event_object, get_transaction_logs
-from .rpc_interface import get_account_balance, get_wait_for_new_block, set_anvil_account_balance
+from .rpc_interface import get_account_balance, set_anvil_account_balance
 from .transactions import smart_contract_read, smart_contract_transact
 from .web3_setup import initialize_web3_with_http_provider
